+++ apis/apis.go
@@ -23,6 +23,7 @@ import (
 	albv1alpha1 "github.com/ionos-cloud/crossplane-provider-ionoscloud/apis/alb/v1alpha1"
 	backupv1alpha1 "github.com/ionos-cloud/crossplane-provider-ionoscloud/apis/backup/v1alpha1"
 	computev1alpha1 "github.com/ionos-cloud/crossplane-provider-ionoscloud/apis/compute/v1alpha1"
+	dataplatformv1alpha1 "github.com/ionos-cloud/crossplane-provider-ionoscloud/apis/dataplatform/v1alpha1"
 	mongov1alpha1 "github.com/ionos-cloud/crossplane-provider-ionoscloud/apis/dbaas/mongo/v1alpha1"
 	postgresv1alpha1 "github.com/ionos-cloud/crossplane-provider-ionoscloud/apis/dbaas/postgres/v1alpha1"
 	k8sv1alpha1 "github.com/ionos-cloud/crossplane-provider-ionoscloud/apis/k8s/v1alpha1"
@@ -39,6 +40,7 @@ func init() {
 		k8sv1alpha1.SchemeBuilder.AddToScheme,
 		albv1alpha1.SchemeBuilder.AddToScheme,
 		backupv1alpha1.SchemeBuilder.AddToScheme,
+		dataplatformv1alpha1.SchemeBuilder.AddToScheme,
 	)
 }