+++ internal/mock/clients/k8s/k8scluster/cluster.go
@@ -38,7 +38,7 @@ func (m *MockClient) EXPECT() *MockClientMockRecorder {
 // CheckDuplicateK8sCluster mocks base method.
 func (m *MockClient) CheckDuplicateK8sCluster(ctx context.Context, clusterName string) (*ionoscloud.KubernetesCluster, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "CheckDuplicateK8sCluster", ctx, clusterName)
+	ret := m.ctrl.Call(m, "CheckDuplicate", ctx, clusterName)
 	ret0, _ := ret[0].(*ionoscloud.KubernetesCluster)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
@@ -47,7 +47,7 @@ func (m *MockClient) CheckDuplicateK8sCluster(ctx context.Context, clusterName s
 // CheckDuplicateK8sCluster indicates an expected call of CheckDuplicateK8sCluster.
 func (mr *MockClientMockRecorder) CheckDuplicateK8sCluster(ctx, clusterName interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CheckDuplicateK8sCluster", reflect.TypeOf((*MockClient)(nil).CheckDuplicateK8sCluster), ctx, clusterName)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CheckDuplicate", reflect.TypeOf((*MockClient)(nil).CheckDuplicateK8sCluster), ctx, clusterName)
 }
 
 // CreateK8sCluster mocks base method.
@@ -69,7 +69,7 @@ func (mr *MockClientMockRecorder) CreateK8sCluster(ctx, cluster interface{}) *go
 // DeleteK8sCluster mocks base method.
 func (m *MockClient) DeleteK8sCluster(ctx context.Context, clusterID string) (*ionoscloud.APIResponse, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "DeleteK8sCluster", ctx, clusterID)
+	ret := m.ctrl.Call(m, "DeleteDataPlatformCluster", ctx, clusterID)
 	ret0, _ := ret[0].(*ionoscloud.APIResponse)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
@@ -78,7 +78,7 @@ func (m *MockClient) DeleteK8sCluster(ctx context.Context, clusterID string) (*i
 // DeleteK8sCluster indicates an expected call of DeleteK8sCluster.
 func (mr *MockClientMockRecorder) DeleteK8sCluster(ctx, clusterID interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DeleteK8sCluster", reflect.TypeOf((*MockClient)(nil).DeleteK8sCluster), ctx, clusterID)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DeleteDataPlatformCluster", reflect.TypeOf((*MockClient)(nil).DeleteK8sCluster), ctx, clusterID)
 }
 
 // GetAPIClient mocks base method.
@@ -98,7 +98,7 @@ func (mr *MockClientMockRecorder) GetAPIClient() *gomock.Call {
 // GetK8sCluster mocks base method.
 func (m *MockClient) GetK8sCluster(ctx context.Context, clusterID string) (ionoscloud.KubernetesCluster, *ionoscloud.APIResponse, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GetK8sCluster", ctx, clusterID)
+	ret := m.ctrl.Call(m, "GetDataplatformClusterById", ctx, clusterID)
 	ret0, _ := ret[0].(ionoscloud.KubernetesCluster)
 	ret1, _ := ret[1].(*ionoscloud.APIResponse)
 	ret2, _ := ret[2].(error)
@@ -108,7 +108,7 @@ func (m *MockClient) GetK8sCluster(ctx context.Context, clusterID string) (ionos
 // GetK8sCluster indicates an expected call of GetK8sCluster.
 func (mr *MockClientMockRecorder) GetK8sCluster(ctx, clusterID interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetK8sCluster", reflect.TypeOf((*MockClient)(nil).GetK8sCluster), ctx, clusterID)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetDataplatformClusterById", reflect.TypeOf((*MockClient)(nil).GetK8sCluster), ctx, clusterID)
 }
 
 // GetK8sClusterID mocks base method.