+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 - **Features**:
 - Add `MongoCluster` crd to support CRUD of MongoDB clusters
 - Add `MongoUser` crd to support CRUD of MongoDB users
+- Add `DataplatformCluster` crd to support CRUD of Dataplatform clusters
 
 ## [1.0.8] (December 2023)
 - **Features**: